+++ HM Studio/src/class_finder.hpp
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <string>
 #include <string_view>
 #include <vector>
 
@@ -10,11 +11,11 @@ private:
 	~Finder() = default;
 
 public:
-	static size_t Find(void* data, size_t dataSize, void* find, size_t findSize, uint32_t start = 0);
-	static void FindAll(void* data, size_t dataSize, void* find, size_t findSize, std::vector<size_t>& output, uint32_t start = 0);
-	static size_t FindBack(void* data, size_t dataSize, void* find, size_t findSize, uint32_t start = 0);
-	static size_t FindFilledBlock(char fill, size_t count, void* data, size_t size, uint32_t start = 0);
-	static size_t FindFilledAlingBlock(char fill, size_t count, void* data, size_t size, uint32_t dtart = 0, size_t align = 0);
+	static size_t Find(void* data, size_t dataSize, void* find, size_t findSize, size_t start = 0);
+	static void FindAll(void* data, size_t dataSize, void* find, size_t findSize, std::vector<size_t>& output, size_t start = 0);
+	static size_t FindBack(void* data, size_t dataSize, void* find, size_t findSize, size_t start = 0);
+	static size_t FindFilledBlock(char fill, size_t count, void* data, size_t size, size_t start = 0);
+	static size_t FindFilledAlingBlock(char fill, size_t count, void* data, size_t size, size_t dtart = 0, size_t align = 0);
 	static bool VerifyFilled(char fill, void* data, size_t size);	
 };
